Revert "Use XML for main editor layout"
This reverts commit 9e977a783fbfabcdc4480a8efd685d457a876592.
